refactor(examples, website): components folder
