+++ examples/03_promise/main.tsx
