+++ examples/04_callserver/components/funcs.ts
