+++ examples/05_mutation/entries.ts
@@ -5,7 +5,7 @@ export default defineEntries(
   async (id) => {
     switch (id) {
       case "App":
-        return import("./src/App.js");
+        return import("./components/App.js");
       default:
         return null;
     }
