+++ examples/05_mutation/index.html
@@ -29,7 +29,7 @@
       <div class="spinner"></div>
     </div>
     <!--/placeholder1-->
-    <script src="/src/index.tsx" async type="module"></script>
+    <script src="/main.tsx" async type="module"></script>
     <!--placeholder2-->
     <!--/placeholder2-->
   </body>
