+++ examples/06_nesting/components/InnerApp.tsx
