+++ examples/07_router/routes/nested.tsx
@@ -1,6 +1,6 @@
 import type { ReactNode } from "react";
 
-import { Counter } from "../src/Counter.js";
+import { Counter } from "../components/Counter.js";
 
 const Nested = ({ children }: { children: ReactNode }) => (
   <div>
