+++ website/main.tsx
@@ -2,7 +2,7 @@ import { StrictMode } from "react";
 import { createRoot } from "react-dom/client";
 import { Router } from "waku/router/client";
 
-import "./index.css";
+import "./main.css";
 
 const root = createRoot(document.getElementById("root")!);
