+++ website/routes/practices/minimal.tsx
@@ -1,4 +1,4 @@
-import { Code, CodeBlock } from "../../src/Code.js";
+import { Code, CodeBlock } from "../../components/Code.js";
 
 const code1 = `import { defineEntries } from "waku/server";
 
@@ -7,7 +7,7 @@ export default defineEntries(
   async (id) => {
     switch (id) {
       case "App":
-        return import("./src/App.js");
+        return import("./components/App.js");
       default:
         return null;
     }
@@ -42,7 +42,7 @@ export default defineEntries(
   async (id) => {
     switch (id) {
       case "App":
-        return import("./src/App.js");
+        return import("./components/App.js");
       default:
         return null;
     }
